use sphinx_simplepdf/parallel_build.py
